requirements.txt
streamlit
openai
python-dotenv
numpy
scikit-learn
sentence-transformers
torch